fix the names of NIST SP 800-53 standards to be the correct names of the publications
"NIST-800-53-rev4" is a kind of fine abbreviation, but it's not the correct title of the publication, which is "NIST SP 800-53 Rev. 4". Maybe the `name` isn't supposed to be used for display purposes, but there's no other place to store the actual name of a standard, so it would make sense to get this field right.

Unfortunately, changing the `name` value within these standards will break OpenControl systems using Compliance Masonry that refer to these standards because Compliance Masonry compares the `standard_key` that appears in components against the standard's `name` field.

Nevertheless, the name is wrong. The OpenControl schema is unclear on the matter and currently suggests by a parallel with the component key that a standard's "key" is its filename if the YAML file doesn't contain a `key` itself. (`key: Key of the component (defaults to the filename if not present)`) One way to look at it, then, is that Compliance Masonry isn't complying with the schema.

Either:

a) Compliance Masonry should be updated to not use the `name` as a `key` and the OpenControl schema should be updated to have the `standard_key` in controls match against a new `key` field rather than `name` in standards, or
b) A new (e.g.) `title` field should be added to the schema to store the actual name of the standards for display purposes.